bloggersapi6lesson fix

+++ src/routers/auth-router.ts
@@ -108,6 +108,7 @@ authRouter.post('/refresh-token', async (req: Request, res: Response) => {
 
 authRouter.post('/logout', async (req: Request, res: Response) => {
   const refreshToken = req.cookies.refreshToken;
+  if (!refreshToken) return res.sendStatus(401);
   const tokenExpire = await jwtService.getUserIdByToken(refreshToken);
   if (tokenExpire === null) return res.sendStatus(401);
   const result = await tokenCollections.deleteOne({ refreshToken: refreshToken });
